moved infinite logic to active_items controller | made common controllers for active_items and requested_items
