+++ app/routes/orders/active_items.js
@@ -3,6 +3,7 @@ import detail from "./detail";
 export default detail.extend({
   async setupController(controller, model = {}) {
     await this._super(controller, model);
+    controller.set("ordersPkgLength", 0);
     controller.on();
   },
