+++ app/routes/orders/requested_items.js
@@ -1,3 +1,5 @@
 import detail from "./detail";
 
-export default detail.extend({});
+export default detail.extend({
+  controllerName: "orders/active_items"
+});
